+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Unmocking using `async fn` when the trait fn has an `-> impl Future` signature.
 
 ## [0.6.0] - 2024-03-25
 ### Changed